project/backend/pyproject.toml
[tool.poetry]
name = "bioinformatics-visualizer"
version = "0.1.0"
description = "Bioinformatics data visualization with LLM integration"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
biopython = "^1.83"
pydantic = "^2.6.0"
python-multipart = "^0.0.7"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
pymongo = "^4.6.0"
httpx = "^0.27.0"
llama-cpp-python = "^0.2.0"
transformers = "^4.38.0"
torch = "^2.2.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"